apps/trading/pyproject.toml
[project]
name = "trading"
version = "0.1.0"
description = "Backtesting framework for Kalshi and Alpaca trading strategies with risk management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.6.0",
    "alpaca-py>=0.34.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "ruff>=0.12.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trading"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/trading/tests"]
